select_mesh_proxy working reliably
